requirements.txt
absl-py==0.4.0
astor==0.7.1
certifi==2018.11.29
cffi==1.12.1
cmake==3.12.0
dlib==19.15.0
face-recognition==1.2.3
face-recognition-models==0.3.0
ffmpy==0.2.2
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
Keras==2.2.0
Markdown==2.6.11
munch==2.3.2
numpy==1.16.2
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==3.4.1.15
pathlib==1.0.1
Pillow==5.4.1
pretrainedmodels==0.7.4
protobuf==3.6.1
pycparser==2.19
scandir==1.7
six==1.12.0
termcolor==1.1.0
torch==1.0.1.post2
torchvision==0.2.1
tqdm==4.25.0